requirements.txt
Django==1.11.6
django-admin-sortable==2.1.1
django-appconf==1.0.2
django-imagekit==4.0.1
django-scribbler==0.8.0
gunicorn==19.7.1
olefile==0.44
pilkit==2.0
Pillow==4.3.0
pytz==2017.2
six==1.11.0
wheel==0.26.0
